% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/settings.R
\name{theme_certara}
\alias{theme_certara}
\title{A ggplot2 theme for Certara.}
\usage{
theme_certara(
  base_size = 11,
  base_family = "",
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  grid = c("none", "horizontal", "both")
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{grid}{Which grid lines should appear? Horizontal only, both horizontal and vertical, or none (default).
\code{\link[ggplot2]{continuous_scale}()}.}
}
\value{
An object of class \code{\link[ggplot2]{theme}()}.
}
\description{
A ggplot2 theme for Certara.
}
\details{
There are 3 variants of the theme: no grid
\code{theme_certara()}, full grid \code{theme_certara(grid = "both")}, and
horizontal grid lines only \code{theme_certara(grid = "horizontal")}.
}
