% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_styling.R
\name{write_code}
\alias{write_code}
\title{Write code to R script from tagged diagnostics}
\usage{
write_code(tagged, file)
}
\arguments{
\item{tagged}{List of tagged objects from returned from \code{VPCResultsUI()}.}

\item{file}{Character specifying path of output file. If missing, it will be saved as \code{code.R} in working directory.}
}
\value{
Returns \code{NULL} after writing to \code{file}.
}
\description{
Use this function to write code to R script from diagnostics tagged in Certara's VPC Results Shiny Application.
}
\examples{
if (interactive()) {
 tagged_diagnostics <-
   vpcResultsUI(tidyvpc::obs_data, tidyvpc::sim_data)

 write_code(tagged_diagnostics, "tagged_vpc.R")
}

}
